Fix arrivals dtype

+++ stochastic_matching/newsimulator/arrivals.py
@@ -36,7 +36,7 @@ def create_prob_alias(mu):
     else:
         cmu = mu
     n = len(cmu)
-    alias = np.zeros(n, dtype=np.int32)
+    alias = np.zeros(n, dtype=int)
     prob = np.zeros(n)
     # noinspection PyUnresolvedReferences
     normalized_intensities = cmu * n / np.sum(cmu)
